Fix MISSING_HEADER_DECLARATION for func_8006417C_64D7C
- Added func_8006417C_64D7C declaration to include/displaylist.h
- Removed extern declaration from src/46080.c
- Updated function signature in src/displaylist.c from u16 to s32 for arg0 to match caller requirements
- Added cast to DisplayListObject* in caller to satisfy type requirements

+++ src/displaylist.c
@@ -454,7 +454,7 @@ void func_8006411C_64D1C(DisplayListObject *arg0) {
     }
 }
 
-void func_8006417C_64D7C(u16 arg0, DisplayListObject *arg1) {
+void func_8006417C_64D7C(s32 arg0, DisplayListObject *arg1) {
     arg1->unk30 = 0;
 
     if (arg1->unk20->opaqueDisplayList != NULL) {
